split out logic to catch netflix subtitles
