+++ background/types/CatchNetflixSubtitlesRequest.ts
@@ -1,4 +1,4 @@
-export type GeminiBatchRequestBody = {
+export type CatchNetflixSubtitlesRequest = {
     message: {
         url: string
         response: any
