+++ background/types/CatchNetflixSubtitlesResponse.ts
@@ -1,0 +1,13 @@
+export type CatchNetflixSubtitlesResponse =
+    | {
+          netflix_sentences: Array<string>
+          error?: never
+      }
+    | {
+          error: string
+          netflix_sentences?: never
+      }
+    | {
+          error?: never
+          netflix_sentences?: never
+      }
