--- background/types/GeminiBatchRequestResponse.ts
@@ -1,9 +1,0 @@
-export type GeminiBatchRequestResponse =
-    | {
-          translatedPhrases: Record<string, string>
-          error?: never
-      }
-    | {
-          error: string
-          translatedPhrases?: never
-      }
